@@ -1,7 +1,7 @@
 [package]
 name = "blockyspot"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 authors = ["BlockySpot Team"]
 description = "A modifiable Spotify Connect client using librespot"
 
